move tests for load() into their own file
